bts: Use uint8_t instead of int for ccch_load_ind_thresh
The range used for this variable is 0-100, so no need to use a signed
integer. Morever, uint8_it is enough for the range.

+++ include/osmocom/bsc/bts.h
@@ -472,7 +472,7 @@ struct gsm_bts {
 	} gprs;
 
 	/* threshold (in percent) when BTS shall send CCCH LOAD IND */
-	int ccch_load_ind_thresh;
+	uint8_t ccch_load_ind_thresh;
 
 	/* RACH NM values */
 	int rach_b_thresh;
